[dmctl.cc] Stop doing substring abbreviations for daemon ids.
It was a good idea in theory but in practice it turns out to be not be a
good win (I don't think I've ever used it) and it causes problems when you
have names like:

  greenfelt
  greenfelt-html

In that case "dmctl greenfelt restart" is ambiguous so you need the fully
qualified name to talk to that daemon.

A better overall solution would be to add bash-completion.