Use a single command socket for all users.
We now authorize users by using the LOCAL_PEERCRED socket option. This
should be equivalent (security-wise) to the old method of using file system
permissions.

This moves all the socket serving/connecting code out of user.cc and into
daemon-manager.cc (for the server), dmctl.cc (for the client), and
command-sock.cc (for the shared sock_addr stuff).

Closes #2